Add `config/balancer` package with popular balancers + remov `config.BalancerConfig` struct + refactor internal balancer package
