+++ internal/driver/cluster/balancer/conn/entry/entry.go
@@ -1,20 +1,20 @@
 package entry
 
 import (
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/conn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/conn/info"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/iface"
 )
 
 // Entry represents inserted into the cluster connection.
 type Entry struct {
 	Conn   conn.Conn
-	Handle balancer.Element
+	Handle iface.Element
 
 	Info info.Info
 }
 
-func (c *Entry) InsertInto(b balancer.Balancer) {
+func (c *Entry) InsertInto(b iface.Balancer) {
 	if c.Handle != nil {
 		panic("ydb: Handle already exists")
 	}
@@ -27,7 +27,7 @@ func (c *Entry) InsertInto(b balancer.Balancer) {
 	}
 }
 
-func (c *Entry) RemoveFrom(b balancer.Balancer) {
+func (c *Entry) RemoveFrom(b iface.Balancer) {
 	if c.Handle == nil {
 		panic("ydb: no Handle to remove from balancer")
 	}
