+++ internal/driver/cluster/cluster.go
@@ -12,11 +12,11 @@ import (
 	"google.golang.org/grpc"
 
 	public "github.com/ydb-platform/ydb-go-sdk/v3/cluster"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/conn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/conn/endpoint"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/conn/entry"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/conn/info"
+	ibalancer "github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/iface"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/balancer/state"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver/cluster/repeater"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/wg"
@@ -36,12 +36,19 @@ var (
 
 	// ErrUnknownEndpoint returned when no connections left in cluster.
 	ErrUnknownEndpoint = errors.New("unknown endpoint")
+
+	// ErrNilBalancerElement returned when requested on a nil Balancer element.
+	ErrNilBalancerElement = errors.New("nil balancer element")
+	// ErrUnknownBalancerElement returned when requested on a unknown Balancer element.
+	ErrUnknownBalancerElement = errors.New("unknown balancer element")
+	// ErrUnknownTypeOfBalancerElement returned when requested on a unknown types of Balancer element.
+	ErrUnknownTypeOfBalancerElement = errors.New("unknown types of balancer element")
 )
 
 type cluster struct {
 	trace    trace.Driver
 	dial     func(context.Context, string) (*grpc.ClientConn, error)
-	balancer balancer.Balancer
+	balancer ibalancer.Balancer
 	explorer repeater.Repeater
 
 	index     map[string]entry.Entry
@@ -73,7 +80,7 @@ type Cluster interface {
 func New(
 	trace trace.Driver,
 	dial func(context.Context, string) (*grpc.ClientConn, error),
-	balancer balancer.Balancer,
+	balancer ibalancer.Balancer,
 ) Cluster {
 	return &cluster{
 		trace:     trace,
@@ -300,10 +307,10 @@ func (c *cluster) Pessimize(ctx context.Context, e endpoint.Endpoint) (err error
 		return fmt.Errorf("cluster: pessimize failed: %w", ErrUnknownEndpoint)
 	}
 	if entry.Handle == nil {
-		return fmt.Errorf("cluster: pessimize failed: %w", balancer.ErrNilBalancerElement)
+		return fmt.Errorf("cluster: pessimize failed: %w", ErrNilBalancerElement)
 	}
 	if !c.balancer.Contains(entry.Handle) {
-		return fmt.Errorf("cluster: pessimize failed: %w", balancer.ErrUnknownBalancerElement)
+		return fmt.Errorf("cluster: pessimize failed: %w", ErrUnknownBalancerElement)
 	}
 	entry.Conn.SetState(ctx, state.Banned)
 	if c.explorer != nil {
